+++ src/design/plone/contenttypes/tests/test_ct_luogo.py
@@ -35,6 +35,7 @@ class TestLuogo(unittest.TestCase):
                 "plone.app.dexterity.behaviors.id.IShortName",
                 "plone.app.dexterity.behaviors.metadata.IBasic",
                 "plone.app.dexterity.behaviors.metadata.ICategorization",
+                "plone.excludefromnavigation",
                 "plone.relateditems",
                 "plone.leadimage",
                 "volto.preview_image",